angrylion HLE: convert the entire geometry/emit path to integer fixed-point
The angrylion HLE geometry frontend, the clip->screen bridge, and the RDP
triangle-command encoder all computed in float (a GLideN64-ism). The real RSP
is integer-only: every coordinate, color, texcoord, depth and slope is an
s15.16 (or s.15 / 0..255) fixed-point quantity. Carrying the pipeline in float
and converting to fixed only at emit time introduced rounding that did not
match the LLE path, most visibly in the depth (Z) plane coefficients.

Remove all floating point from the HLE path:
 - EmitVertex / GSPVertex / BridgeViewport now hold s15.16 (and 0..255 color,
   S10.5 texel) integer fields; the lossy float clip-space mirror is gone.
 - clip_to_emit does the perspective divide with the RSP fixed reciprocal
   (rsp_recip_div) and the viewport map entirely in integer, no float round
   trip.
 - emit_*_triangle solve the edge, color, texture/W and Z planes with 64-bit
   integer cross-products and integer division (solve_plane), packing the RSP
   coefficient words directly.
 - gsp_vertex lighting (normal transform, normalize, n.L) is integer: s8
   normal x s15.16 modelview in a 64-bit accumulator, integer sqrt for the
   normalize, 0..255 color; light state stored as integers (rgb 0..255, dir
   s.15).
 - viewport, light load, cull and degenerate-reject all integer.
 - the Z coefficient is the screen z scaled to the RDP sub-precision domain
   (screen_z * 32, matching the cxd4 LLE z base bit-for-bit).

No float code remains in the HLE path (only comments mention it). C89/MSVC
clean. On Hyrule Field this moves the whole-frame divergence from the cxd4 LLE
oracle from ~47% to ~32%; the residual is the Z-plane slope derivation, tracked
separately.

+++ mupen64plus-video-angrylion/rdp_emit.h
@@ -13,11 +13,16 @@ typedef signed   __int32 rdp_emit_int32_t;
 extern "C" {
 #endif
 
+/* One projected vertex in the RSP's native fixed-point forms (no float). The
+ * RSP/RDP are integer-only. x,y are s15.16 screen coordinates; z is s15.16
+ * screen depth (0..0x3ffff range in the integer part path); w is the RDP
+ * inverse-w coefficient (perspNorm-scaled 1/w); r,g,b,a are 0..255 color held
+ * as s15.16; s,t are the S10.5 texel coordinate held as s15.16. */
 typedef struct EmitVertex
 {
-    float x, y, z, w;
-    float r, g, b, a;
-    float s, t;
+    int32_t x, y, z, w;
+    int32_t r, g, b, a;
+    int32_t s, t;
 } EmitVertex;
 
 int emit_shaded_triangle(int32_t *ew, const EmitVertex *va,
